Refactor - Competitors Page
- Split Process Results
