+++ backend/routes/__init__.py
@@ -11,6 +11,7 @@ __all__ = [
     "leagues",
     "leagueEvents",
     "leagueResults",
+    "leagueCompetitors",
     "results",
     "resultsManual",
     "resultsRecalculate",
